LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-pathLength.tentative-expected.txt


FAIL SMIL from-to animation of the 'pathLength' attribute assert_equals: expected "125" but got ""
